Abstract starting application subservices
Creates a Services abstraction which handles running and shutting down of services that the application will boot.

The Services abstraction adds an additional (as yet unused) feature which allows for blocking services to be run in a goroutine. This allows us to push responsibility for handling goroutines up to caller which should make testing easier.